update nicer coe in general
