+++ examples/dev/singleTop/fung.g96
@@ -1,23 +1,29 @@
 TITLE
-	 >>> Generated with python lib function_libs utilities. (riniker group)	 >>> line_seperator: '\n'	 field_seperator: '\t'	 comments_char: '#'
+	 >>> Generated with python lib function_libs utilities. (riniker group)	 >>> line_seperator: '\n'	 field_seperator: '\t'	 comments_char: '#'	 >>> Generated with python lib function_libs utilities. (riniker group)	 >>> line_seperator: '\n'	 field_seperator: '\t'	 comments_char: '#'
 	 >>> Generated with python lib function_libs utilities. (riniker group)
 	 >>> line_seperator: '\n'	 field_seperator: '\t'	 comments_char: '#'
 END
 POSITION
 # 	 
-    1 F378  C5         1   -0.213727417    0.354335876   -0.003240672
-    1 F378  O1         2   -0.093019160    0.279001792    0.000243936
-    1 F378  C4         3   -0.102939305    0.141918380    0.000711725
-    1 F378  C3         4    0.018250526    0.071394519    0.000004694
-    1 F378  H4         5    0.111304909    0.127475534   -0.000850612
-    1 F378  C2         6    0.019387384   -0.067653748   -0.001712013
-    1 F378  H3         7    0.114627096   -0.120425292   -0.002634783
-    1 F378  C1         8   -0.100731452   -0.141029430   -0.002129426
-    1 F378  N1         9   -0.099284927   -0.281761402    0.003451761
-    1 F378  H1        10   -0.182330588   -0.324460779   -0.036355167
-    1 F378  H2        11   -0.016782612   -0.323047736   -0.038974488
-    1 F378  C7        12   -0.221332678   -0.069621200   -0.001120014
-    1 F378  H9        13   -0.315639720   -0.123849744   -0.000763903
-    1 F378  C6        14   -0.223083517    0.070253457    0.000796961
-    1 F378  H8        15   -0.318807150    0.121347377    0.001516279
+    1 F378  C          1   -0.213738224    0.354296869   -0.003149189
+    1 F378  C          2   -0.093034159    0.279002559    0.000232480
+    1 F378  C          3   -0.102927162    0.141901993    0.000715450
+    1 F378  C          4    0.018276367    0.071408239   -0.000002296
+    1 F378  C          5    0.111273454    0.127513078   -0.000901660
+    1 F378  C          6    0.019383309   -0.067691484   -0.001720367
+    1 F378  C          7    0.114585920   -0.120386532   -0.002633961
+    1 F378  C          8   -0.100712998   -0.140997530   -0.002121299
+    1 F378  C          9   -0.099305450   -0.281798182    0.003460389
+    1 F378  C         10   -0.182306122   -0.324497236   -0.036439226
+    1 F378  C         11   -0.016806384   -0.322988495   -0.039050813
+    1 F378  C         12   -0.221316551   -0.069603785   -0.001103479
+    1 F378  C         13   -0.315613778   -0.123808612   -0.000803777
+    1 F378  C         14   -0.223123519    0.070295877    0.000814745
+    1 F378  C         15   -0.318826059    0.121290926    0.001528154
+    1 F378  C         27    0.148900000    0.248300000    0.000500000
+    1 F378  C         28    0.266900000    0.070700000   -0.000000000
+    1 F378  C         29    0.361200000    0.124900000    0.000300000
+    1 F378  C         30    0.266900000   -0.070700000   -0.000000000
+    1 F378  C         31    0.361200000   -0.124900000    0.000300000
+    1 F378  C         33    0.148900000   -0.248300000    0.000500000
 END
